global_timer: fixed for both autonomous and joystick mode
In joystick control mode global_timer incorrectly started to count
as soon as 'play' was pressed.

This is because 'self._processing' variable in RemoteController class
was incorrectly set to True. Same flag has to be controlled more
precisely with help of additional callback, on_joystick_action that
could tell when its time to start counting the global_timer