remote_controller.py: control_button_pressed converted to class
Logic not changed.

control_button_pressed is using numeric values with are unreadable.
Same numeric values are expected to be used in different py files, that
makes it hard to debug.
Also there is an implicit fetch and clear semantics.
All above usecases opt for using a class with well defined and limited
API.