Fixed script variable characteristic response
Function 'update_script_variable' is used to generate response
for a BLE characteristic, writing out 4 slots for script variables
The protocol also implies having a mask byte that shows which of the
slots is set at least once. Example - background script has not started
to run, so variable is never set. In that case bit in the mask will be
zero.

Previous code did not work