Implemented validate configuration characteristic
Added new characterstic "Current Port Configuration" according
to updated spec.
UUID=AD635567-07A7-4C8A-8765-D504DAC7C86F

This is used to check that user configuration is valid by asking
firmware if peripherals, specified in configuration are physically
connected to the ports.

Feature is not completely defined, as many things are not working
yet on the firmware/hardware side