TestMotorOnPort: workaround for I2C communication problem
TestMotorOnPort is a firmware command that runs asynchronously, that is
it sets up for a long run and in first response to OpStart command sends
'Pending' status. It is the job of the caller application to poll with
OpGetResult operation until the firmware returns Ok or Error instead of
Pending.
Right now communication gets broken if OpGetResult is called without any
time interval. This should be fixed on a firmware level, but as a quick
fix robot code should wait more before calling next GetResult.