TestMotorOnPort: support new robot fw with parametrized motor power
Motor detection function works in a way that motor is driven by motor
driver by a small amount of current and in the same time ADC senses
how much current is going through the motor winding. If motor is
attached, ADC will sense increased current on the corresponding motor
driver current pin.
Excessive motor movement is not welcome, because motor detection is used
in config validation and during this robot should stay in the same
position if possible.
Less motor power will make motor movement more subtle, but also might be
invisible to ADC, and leads to detection failure.
It is useful to fine-tune this value outside of the firmware as an
argument to motor detection request.