Drivetrain: fixed issue with unfinished motor program
Most of the block programs are expected to have deteministic lifetime,
these programs should as soon as the goal is reached, ex. drivetrain program
to turn 90 degrees CW should end as soon as 90 degrees goal is reached.
In the end of the program awaiter.finished is called, this runs the
callbacks, stopping the program, but self._controller was still working
after all the finishing callbacks are done. Also self._controller is
present in _on_motor_status_changed callback, that registers any motor
statuses, coming from chip with firmware. While receiving statuses is
ok, because the drivetrain is setup, running controller.update inside of
that is not ok. Controller should have the lifetime of block program, so
it should be cleaned up also within the scope of awaiter.finish function
call